fix windows compilation
git-svn-id: http://svn.cern.ch/guest/lhcb/Lbcom/trunk@124993 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: b1733d3bb2b780effe868999dc30d61759771e95

+++ Lbcom/Calo/CaloTools/src/CaloHypoEstimator.cpp
@@ -251,7 +251,7 @@ bool CaloHypoEstimator::estimator(const LHCb::CaloHypo* hypo){
         
         if ( 0 == *id || !isPrs ( *id ) )continue;
         LHCb::CaloCellID id2 = (*id)->cellID();
-        if( abs((*id)->cellID().row() - idseed.row())<2 && abs( (*id)->cellID().col() - idseed.col())<2 ){
+        if( abs((int)(*id)->cellID().row() - (int)idseed.row())<2 && abs( (int)(*id)->cellID().col() - (int)idseed.col())<2 ){
           // Build sum1 and sum9
           //
           sum9 += (*id)->e();
